requirements.txt
fastapi[all]==0.115.11
sqlalchemy==2.0.39
passlib[bcrypt]==1.7.4
psycopg2==2.9.10
sqlmodel==0.0.24
pyjwt==2.10.1